refactor(coupon-kafka-consumer): update kafka consumer use failure recorder
